server/scripts/requirements.txt
pandas==2.2.2
scikit-learn==1.5.1
matplotlib==3.9.1
google-generativeai==0.7.2
numpy==2.0.1
scipy==1.14.0
